created a common tabs component
