+++ src/App.js
@@ -4,6 +4,10 @@ import MainTab from './components/tabs/mainTab'
 import TabsComponent from './components/tabs/tabsComponent'
 import useYoroi from './hooks/yoroiProvider'
 import {CONNECTED, NO_CARDANO} from './utils/connectionStates'
+import Cip30Tab from './components/tabs/subtabs/cip30Tab'
+import Cip95Tab from './components/tabs/subtabs/cip95Tab'
+import Cip95TabTools from './components/tabs/subtabs/cip95ToolsTab'
+import NFTTab from './components/tabs/subtabs/NFTTab'
 
 const App = () => {
   const {connectionState} = useYoroi()
@@ -15,11 +19,34 @@ const App = () => {
     isNotCardanoWallet,
   }
 
+  const data = [
+    {
+      label: 'CIP-30',
+      value: 'cip30',
+      children: <Cip30Tab />,
+    },
+    {
+      label: 'CIP-95',
+      value: 'cip95',
+      children: <Cip95Tab />,
+    },
+    {
+      label: 'CIP-95 Tools',
+      value: 'cip95Tools',
+      children: <Cip95TabTools />,
+    },
+    {
+      label: 'NFTs',
+      value: 'nfts',
+      children: <NFTTab />,
+    },
+  ]
+
   return (
     <div className="min-h-screen bg-gray-800">
       <AccessButton />
       <MainTab {...mainTabProps} />
-      <TabsComponent />
+      <TabsComponent tabsData={data} />
     </div>
   )
 }
